[SPARK-XXXXX][SQL] Extract V2 runtime-filter + partition planning into a shared `PushDownUtils.filterAndPlanPartitions` helper
### What changes were proposed in this pull request?

Lift the body of `BatchScanExec.filteredPartitions` (runtime filter pushdown,
re-planning, and `KeyedPartitioning` validation + `None`-padding) into a new
`PushDownUtils.filterAndPlanPartitions` helper so the logic can be reused by
alternative DataSourceV2 physical scan operators in downstream forks
without duplicating ~60 lines of subtle SPJ-alignment code.

Three mechanical pieces:

1. `PushDownUtils.pushRuntimeFilters` now takes a by-name
   `Option[Seq[PartitionPredicateField]]` instead of `Table`. The only
   thing it needed from `Table` was `table.partitioning` for the
   iterative-pushdown branch; passing the precomputed schema lets callers
   that have a `Scan` but not a `Table` (e.g., a connector that holds the
   partition transforms separately) reuse this code path. By-name avoids
   any cost for scans that don't opt into iterative pushdown.
2. New `getPartitionPredicateSchema(transforms: Array[Transform], output)`
   overload that the existing `(table, output)` overload now delegates to.
3. New `filterAndPlanPartitions(scan, runtimeFilters, partitionPredicateSchema,
   output, outputPartitioning, inputPartitions): Seq[Option[InputPartition]]`
   helper. `BatchScanExec.filteredPartitions` is now a one-line call.

No behavior changes for existing callers. The only caller of
`pushRuntimeFilters` in `apache/spark` was `BatchScanExec.filteredPartitions`,
which is now replaced. The `getPartitionPredicateSchema(relation)` and
`(table, output)` overloads remain for other call sites
(`V2ScanRelationPushDown`, `GroupBasedRowLevelOperationScanPlanning`,
`OptimizeMetadataOnlyDeleteFromTable`).

### Why are the changes needed?

The runtime-filter pushdown pipeline in `BatchScanExec.filteredPartitions`
contains non-trivial logic: V2 predicate translation (DPP + scalar
subqueries via SPARK-56467), iterative `PartitionPredicate` pushdown
(SPARK-55596), `KeyedPartitioning` validation, and `None`-padding to
preserve SPJ key alignment. As more pushdown rounds get added (e.g., the
proposed `SupportsPushDownCatalystRuntimeFiltering` for Catalyst-runtime
filter pushdown), this logic gets harder to keep in sync across
alternative DSv2 physical operators.

Extracting the helper:
- Keeps the partition-pruning pipeline as a single source of truth.
- Lets downstream forks (e.g., the Databricks Photon scan operator) call
  the same helper instead of inlining a subset, avoiding silent drift.
- Decouples `pushRuntimeFilters` from `Table`, so callers that hold a
  `Scan` and a `Transform[]` (but no `Table`) can reuse it.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests cover the helper path end-to-end through `BatchScanExec`:
- `DataSourceV2EnhancedRuntimePartitionFilterSuite` exercises rounds 1
  (translatable V2 predicates) and 2 (iterative `PartitionPredicate`).
- `DataSourceV2SQLSuiteV2Filter.SPARK-56467` exercises scalar-subquery
  partition pruning.

No new tests added; the helper is a direct extraction with no behavior
change.

### Was this patch authored or co-authored using generative AI tooling?

Generated-by: Claude (Anthropic), via Claude Code
